+++ project-base/src/Shopsys/ShopBundle/Model/Newsletter/NewsletterFacade.php
@@ -4,6 +4,7 @@ namespace Shopsys\ShopBundle\Model\Newsletter;
 
 use DateTimeImmutable;
 use Doctrine\ORM\EntityManager;
+use Shopsys\ShopBundle\Form\Admin\QuickSearch\QuickSearchFormData;
 
 class NewsletterFacade
 {
@@ -45,5 +46,35 @@ class NewsletterFacade
     public function getAllEmailsDataIteratorByDomainId($domainId)
     {
         return $this->newsletterRepository->getAllEmailsDataIteratorByDomainId($domainId);
+    }
+
+    /**
+     * @param int $selectedDomainId
+     * @param \Shopsys\ShopBundle\Form\Admin\QuickSearch\QuickSearchFormData $searchData
+     * @return \Doctrine\ORM\QueryBuilder
+     */
+    public function getQueryBuilderForQuickSearch(int $selectedDomainId, QuickSearchFormData $searchData)
+    {
+        return $this->newsletterRepository->getQueryBuilderForQuickSearch($selectedDomainId, $searchData);
+    }
+
+    /**
+     * @param int $id
+     * @return \Shopsys\ShopBundle\Model\Newsletter\NewsletterSubscriber
+     */
+    public function getNewsletterSubscriberById(int $id)
+    {
+        return $this->newsletterRepository->getNewsletterSubscriberById($id);
+    }
+
+    /**
+     * @param int $id
+     */
+    public function deleteById(int $id)
+    {
+        $newsletterSubscriber = $this->getNewsletterSubscriberById($id);
+
+        $this->em->remove($newsletterSubscriber);
+        $this->em->flush();
     }
 }